Add simple input validation in exc 4
Check if entered input is greater than max width or height of src image. Check if method input was correct.

For a more complete validation, method needs to check for sum of starting index and width or height going out of range. Also check for input being smaller than or equal to zero.