Add looking up article by name
Search for a matching article name within the articles list. If found,
display article info, else display that no article with that name was
found in the list.

To meet the exercise requirements, looking up an article by name will
only allow the user to display information about the article. To modify
the article object, users can search for the article by its list index
and accessing the object properties.