Validate new property value
Display an error message in the object property TextArea in case of an
incorrect property type set in the TextField or the TextArea element.

The method convertFieldType is used to cast the content of the child
element to the expected type of the targeted property, after which a
property setter is called. In case of an IllegalArgumentException being
thrown, it is assumed that an incorrect value was provided as the new
property value.