Change validation trigger
Listen for key press from the text container's FXML property. Execute
the validateField method via the 'onKeyTyped' property.

This change simplifies the HelloController code and makes listening for
the change easier. Old implementation listened for a value change of a
TextField element. That solution would need two sets of triggers to
work for both the TextField and TextArea elements. Performing the
validation on key press creates a universal trigger.