Fix query type conversion issue
Revert the last commit to return Object[] type from the SQL query.
Create a separate method for returning all available rooms in a
List<Room> type object.

Changes made in the last commit threw an exception and introduced a
problem with getting hotel rooms in the getAvailableRoomsByHotel
method. The approach had returned the room capacity instead of the
available rooms per hotel location.