Revert command length reading fix
Revert back to checking characters using static indexes of the tmp
array instead of calculating the index based on array data length.

This fixes the previous implementation's problem with mismatched
character indexes in the tmp array. At the time of validating the
characters in code, there was not enough data to properly calculate the
index.