Refactor car options relationships
Recreate the SamochodOpcje table, establishing the connection between
cars in the Samochod table and their options in the Opcje table.

Add a precision to the Cena attributes in the Samochod and Opcje
tables.

The table was reintroduced to align with the project's requirements of
having a multivalued attribute in one of the tables.