Fix the command length calculation
Use the strlen function to get the size of the array.

Add a temporary cmd_len_fill variable to store and fill the command
length in the returning frame.

Null terminate the command array after filling it in the frame_get
function.

Null terminate the result array in the frame_send function.