Dynamically select the date picker components
Select which picker to present to the user based on the selection in
the SelectList component.

Update the backend code to properly process the formData object.

The Google API only allows for specifying the task's due date. Adding a
second date picker component prevents the confusion of the selected
time not getting applied to the newly created task.