Update local database based on remote changes
Add the 'getDatabaseSyncStatus' function to check for changes between
local and remote data. If there are changes in the remote data, update
the locally stored events.

This commit implements the functionality to compare local and remote
event entries, allowing for updates in the local database in the event
of remote changes.