Optimize adding local database event entries
Import the useRef hook to track event objects that differ from the
remote database events. The useRef is used to obtain access to the
objects within the entire scope of the App function.

Update the 'updateLocalData' function body to use the contents of the
new hook to add events.

This commit optimizes the logic for adding events to the local database
by only adding entries from the useRef 'eventsToAdd' array instead of
the entire 'calendarData' events array.