Update ScrollView boundaries in the main view
Extend the container height and remove the padding.

Replace inline styling for the element with a stylesheet property.

This commit updates the ScrollView styling in preparation for the
addition of 'AlertBox.tsx'. The styling of the pop-up alert could not
have been properly implemented with the previous 'App.tsx' component
styling.