Update UI after entry operations
Create the 'fetchRemoteData()' function and export it as
'refetchRemoteData()' to 'App.tsx'.

Expand the dependency array responsible for calling
'updateAndFetchLocalData()' to trigger on a change in database sync
status.

Create a 'withRefetch()' function to refetch calendar data directly
after adding a new entry in 'App.tsx'.

This commit updates the remote calendar data after submitting a new
entry or altering an existing entry, updating the entries in the UI.