Adjust sensor communication delays
Adds a delay between wake-up retries in case of a transmission error in
'AM2320_InitSensorRead()'.

Changes the delay between successful wake-up and sending the read
command to a ~1ms delay in 'AM2320_InitSensorRead()'.

This commit adds a timeout delay on a failed wake-up call and makes the
delay between the wake-up and sending the read command shorter.