Implement reading archived sensor reads
Changes 'main.c' -- adds the 'convert_char_string_to_uint16()' function
to convert the index from the frame to an integer, changes the
'AM2320_SendSensorDataFrame()' parameters to allow for the function to
be used with, and without returning the index in the return frame.

This commit adds reading data from a specific index in the AM2320 data
buffer with the 'READ' command.