Make 'receive_frame()' return data length
Changes 'main.c' -- changes the 'receive_frame()' declaration and logic
to make the function return the amount of collected data.

Changes 'main.c' -- adds the 'data_length' argument to the existing
calls of the 'receive_frame()' function.

This commit lets the user retrieve the amount of received data from the
'receive_frame()' function.