Fix user balance UI element updates
Changes 'SummaryViewModel.cs' -- adds a subscription to changes of the
transaction list that triggers the balance and summary info UI updates.

Changes 'UserRepository.cs' -- adds a method for getting the default
user from the database directly from the method call.

This commit implements working updates on transaction changes for the
'CurrentBalance' UI element on the 'SummaryView' screen.