Fix displaying transactions' category names
Changes 'MainViewModel.cs' -- supplies 'CalendarViewModel' with the
transaction categories' collection.

Changes 'CalendarViewModel.cs' -- adds a transaction category object to
each object in the 'TransactionsFromSelectedDate' collection.

Changes 'TransactionCategoryRepository.cs' -- changes the data type of
the collection returned from the get-all method.

This commit adds binding to the 'Name' property from the
'TransactionCategory' class to the objects in the
'TransactionsFromSelectedDate' collection.