Implement submitting alerts to the database
Changes 'MainViewModel.cs' -- initializes and provides the alarm
repository to the summary view.

Changes 'SummaryViewModel.cs' -- modifies the constructor to expect an
alarm repository, to load and add the alarms.

Changes 'SummaryView.xaml' -- adds the alarm box alongside the monthly
summary items.

Changes 'SummaryView.xaml.cs' -- updates the code-behind to allow
binding the view model properties to the pop-up alert box.

This commit updates the codebase to allow the user to add an alert to
the database.