Prevent running queries during application loading
Changes 'MainViewModel.cs' -- supplies a loading state variable to the
SummaryViewModel.

Changes 'SummaryViewModel.cs' -- adds checks for the application state,
to prevent the services from running during the loading period.

This commit fixes the problem with the user balance and services like
alert or spending service being run on every loaded transaction when
launching the application.